requirements.txt
python-tkdnd
validators
requests
Pillow
pytesseract